Moved classroom-detail file into its own folder and made notes section render conditionally.
