+++ src/scenes/cohort/classroom/classroom-detail.tsx
@@ -1,33 +1,20 @@
 import * as React from 'react';
 import { connect } from 'react-redux';
-import { RootReducerState } from '../../state/reducers';
-import { ModuleComponent } from '../../components/content-module/content-module';
+import { RootReducerState } from 'state/reducers';
+import { ModuleComponent } from '../../../components/content-module/content-module';
 import { Route, RouteComponentProps, Switch } from 'react-router';
-// import { CalendarComponent } from '../../components/calendar/calendar';
-// import { AddClassroomFormModal } from '../../components/calendar/add-classroom-form-modal';
-// import {
-//   toggleCohortClassroomDialog,
-//   updateClassroomInEdit,
-//   saveClassroomToCohort
-// } from '../../state/actions/cohortActions';
-import { convertObjectToValuesArray, formatDateStringWithoutTime, formatAmPmTime } from '../../helpers';
-// import { ModuleListItem } from '../home/module-list/module-list-item/module-list-item.component';
-// import { Card, CardContent, Typography } from '@material-ui/core';
+import { formatDateStringWithoutTime, formatAmPmTime } from 'helpers';
 import { Link } from 'react-router-dom';
-
-
 
 export const ClassroomDetailComponent =
   ({
     selectedClassroom,
-    // cohortModules,
     selectedCohort,
     classroomModules,
-
    }) => (
     <React.Fragment>
 
-        {selectedCohort &&
+        { selectedCohort &&
             <Link to={`/cohorts/${selectedCohort.id}`} style={{ textDecoration: 'none' }}>
                 <button className="bg-light-green bn ph3 pv2 pointer">Back to cohort page</button>
             </Link>
@@ -37,20 +24,25 @@ export const ClassroomDetailComponent =
 
             <p>
                 <span className="b">Class Date: </span>
-                {selectedClassroom.day ? formatDateStringWithoutTime(new Date(selectedClassroom.day)) : ""}
+                { selectedClassroom.day ? formatDateStringWithoutTime(new Date(selectedClassroom.day)) : "" }
             </p>
             <p>
                 <span className="b">Start Time: </span>
-                {selectedClassroom.startTime ? formatAmPmTime(selectedClassroom.startTime) : ""}
+                { selectedClassroom.startTime ? formatAmPmTime(selectedClassroom.startTime) : "" }
             </p>
             <p>
                 <span className="b">End Time: </span>
-                {selectedClassroom.endTime ? formatAmPmTime(selectedClassroom.endTime) : ""}
+                { selectedClassroom.endTime ? formatAmPmTime(selectedClassroom.endTime) : "" }
             </p>
-            <p>
-                <span className="b">Additional Notes: </span>
-                {selectedClassroom.notes ? selectedClassroom.notes : ""}
-            </p>
+
+            { selectedClassroom.notes &&
+                (
+                    <p>
+                        <span className="b">Additional Notes: </span>
+                        { selectedClassroom.notes }
+                    </p>
+                )
+            }
 
             <div className="modules-container">
             <p className="b">Modules in this class:</p>
@@ -71,7 +63,8 @@ export const ClassroomDetailComponent =
 
 export const ClassroomStateful = connect((state: RootReducerState, ownProps: RouteComponentProps<any>) => {
 
-    const classroomById = (state.cohort.allCohorts[ownProps.match.params.name] && state.cohort.allCohorts[ownProps.match.params.name].classrooms) ? convertObjectToValuesArray(state.cohort.allCohorts[ownProps.match.params.name].classrooms).find((classroom) => classroom.id === ownProps.match.params.id) : [];
+    const classroomById = (state.cohort.allCohorts[ownProps.match.params.name] && state.cohort.allCohorts[ownProps.match.params.name].classrooms) ? 
+    state.cohort.allCohorts[ownProps.match.params.name].classrooms[ownProps.match.params.id] : [];
 
     const classroomModuleIds = classroomById.modules || [];
 
@@ -82,7 +75,6 @@ export const ClassroomStateful = connect((state: RootReducerState, ownProps: Rou
     ...ownProps,
     selectedCohort: state.cohort.allCohorts[ownProps.match.params.name],
     selectedClassroom: classroomById,
-    // cohortModules: cohortModules,
     classroomModules: cohortModules ? cohortModules.filter(module => classroomModuleIds.includes(module.id)) : [],
   };
 }, {
